+++ OgreMain/src/Emscripten/OgreTimer.cpp
@@ -50,27 +50,27 @@ void Timer::reset()
 }
 
 //--------------------------------------------------------------------------------//
-unsigned long Timer::getMilliseconds()
+unsigned long long Timer::getMilliseconds()
 {
     return emscripten_get_now() - start;
 }
 
 //--------------------------------------------------------------------------------//
-unsigned long Timer::getMicroseconds()
+unsigned long long Timer::getMicroseconds()
 {
     return (emscripten_get_now() - start) * 1000.0;
 }
 
 //-- Common Across All Timers ----------------------------------------------------//
-unsigned long Timer::getMillisecondsCPU()
+unsigned long long Timer::getMillisecondsCPU()
 {
     clock_t newClock = clock();
-    return (unsigned long)((float)(newClock-zeroClock) / ((float)CLOCKS_PER_SEC/1000.0)) ;
+    return (unsigned long long)((float)(newClock-zeroClock) / ((float)CLOCKS_PER_SEC/1000.0)) ;
 }
 
 //-- Common Across All Timers ----------------------------------------------------//
-unsigned long Timer::getMicrosecondsCPU()
+unsigned long long Timer::getMicrosecondsCPU()
 {
     clock_t newClock = clock();
-    return (unsigned long)((float)(newClock-zeroClock) / ((float)CLOCKS_PER_SEC/1000000.0)) ;
+    return (unsigned long long)((float)(newClock-zeroClock) / ((float)CLOCKS_PER_SEC/1000000.0)) ;
 }
